JwtOptions and its builder were moved out to BMJ.Authenticator.Adapter project
